Remove sumbool/decidability section.
This DEFINITELY breaks the code, as one of these functions is used
twice in other proofs. However, this will be rectified in the next
few commits.

WARNING: this is part of a series of changes incrementally committed
and was not tested to compile successfully. In fact, this WILL NOT
compile successfully.