Switch from assumptions being Prop-valued to Type-valued.
This means the argument to `by_assumption` is _allowed_ (but does
not have to) be a datatype i.e. its terms are not indistinguishable.

This greatly simplifies e.g. the Deduction Theorem by allowing the
combination of two different sets of assumptions to be their `sum`
rather than their `or`.

Along with this,
+ make some changes to argument/section/etc. structure of file
+ add new notations ⊔, ⊆' (however, note that the latter doesn't work currently)
+ Update existing code to use ⊔ instead of ∪, replace
  `or`/`introl`/`intror`/etc. with `sum`/`inl`/`inr`/etc. and so on
+ rewrite the transitivity lemmas without complicated `dec`s,
  and only two verbose `Fixpoint` definitions (ideal is one with rest in
  terms of that one)
+ rewrite the deduction theorem without any `dec` or descending to `Prop`

WARNING: this is part of a series of changes incrementally committed
and was not tested to compile successfully.