Reverse the order of arguments to modus ponens.
Reverse the order of arguments to `modus_ponens` as well as
`modus_ponens_under_imp`, so the proof of the implication comes first,
_then_ the proof of the hypothesis of the implication.
Additionally, modify lemmas using them accordingly. In the particular
case of `exfalso`, modify the proof itself.

It is much more common to apply `modus_ponens` to just the implication
argument (in which it states that Proof is sort of a 'homomorphism'
wrt implication), or for the implication argument to be small with
the hypothesis proof argument being the rest of the proof.

Thus, for reasons of partial application and `apply` tactic-like usage,
it is better for the implication argument to come first.

Among the lemmas using `modus_ponens` or `modus_ponens_under_imp` which
were modified, note in particular the simplification of the second half
of `or_spec` of `LindenbaumTarksiAlgebra`, and to a lesser extent,
`interchange_hypotheses`.