Add `deduction_theorem'` and `red_by_dt` tactic and other useful tactics.
`deduction_theorem'` allows the use of the deduction theorem to be
"turned off" at a certain level, only using the deduction theorem's
transformation for a certain subtree of the proof tree.  This allows
going back and forth between having the hypothesis as an assumption and
an implication in the goal.  This is practically usable through the
tactic `reducing_deduction_theorem`, whose user interface is the tactic
notation `red_by_dt`.

Additionally, the tactic functionality for decomposing assumptions using
⊔ and eq has been extracted into `detect_assumption` (for constructing)
and `intro_assumption` (for introducing and decomposing an assumption).

Moreover, remove several `Section` commands that don't introduce new
variables, to prevent accidentally restricting tactics/tactic notations
to sections (this having happened with`red_by_dt`/
`reducing_deduction_theorem`).

Heavily refactor a large part of the code to use these new features.
Moreover, almost complete `LindenbaumTarksiAlgebra`.