Refactor src/Lattices.v, (mainly) turning BooleanAlgebra into a bundled class.
Specifically, add `isBooleanAlgebra` class with convenient, i.e. easy
to prove, required fields and the previous, stronger, properties as
"manifest fields" i.e. proved directly in the class definition once
and for all.

Make BooleanAlgebra a structure bundling an `isBooleanAlgebra`
instance.

Also:
+ Rename boolean algebra operations to `join`, `meet`, `bot`, `top`,
  `complement`; and other naming changes.
+ Replace refl/rfl idiom with multiple `Arguments` idiom.
+ Change boolean algebra notations to Unicode versions and remove
  `Local` annotations.  (Also stop importing these notations in
  src/propositional_logic.v.)
+ Change `filter` (now `Filter`) to allow improper filters and add
  a predicate `filter_proper` and type `ProperFilter` for proper
  filters.

WARNING: this is part of a series of changes incrementally committed
and was not tested to compile successfully (although it probably
does).