Complete definition of `LindenbaumTarksiAlgebra`.
This requires refactoring of `isBooleanAlgebra` as well to take
a `Preordered` "unbundled" as a type, relation and instance of
`PreOrder` due to mysterious type inference errors related to
expecting `carrier {| le := provable_le |}` and getting `Proposition`.

The proof of conjunction distributing over disjunction requires the
`join_le_of_both_le` property of disjunction for a _different_ set
of assumptions, which meant Γ cannot be held fixed over the entire
definition. Instead, prove a bunch of lemmas which are basically the
fields with Γ not held fixed, then construct the algebra.