Restrict `intro_assumption` to using identifiers to name the hypothesis.
Do this by replacing the use of `destruct` in branches of the
recursive destruction to `rename`.

This fixes the problem caused by using `destruct` that it complains
when (apparently) the hypothesis to be destructed is not of an
inductive type, for example of type `F x` for variables `F` and `x`.