Give up on many-sorted FOL (at least, for now) and define substitution.
Thus, change
+ heterolist into Array
+ occurrence into BNat
+ `vararg` functions to take a nat and a type rather than a list
  (indexing) types
+ Additionally, remove the extra `types` parameter of the signature and
  all the `type` arguments
+ etc.

Additionally, complete definition of substitution by an array of
values for the full context, rather than for a single variable. This
eliminates the huge struggle of trying to define substitution of a
single variable by a single term, involving trying to prove decidable
equality on `BNat`, etc.

This idea came from a few resources on first-order logic.