Delete _CoqProject as it seems to be unnecessary.
Perhaps it will be useful if this project depends on projects
outside the standard library, or needs to specify make file
instructions or something else.  But since modules in the same
folder can be required with no prefix anyway, and all the .v
files are in that folder, it's unnecessary now.